# backend-api/swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/users":
    post:
      summary: Creates a new user
      tags:
      - Users
      parameters: []
      responses:
        '201':
          description: User created
        '422':
          description: Invalid request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                identification_number:
                  type: integer
                balance:
                  type: number
                pin:
                  type: integer
                face_image_url:
                  type: string
              required:
              - name
              - identification_number
              - balance
              - pin
              - face_image_url
    get:
      summary: Retrieves all users
      tags:
      - Users
      responses:
        '200':
          description: Users retrieved
  "/users/{id}":
    get:
      summary: Retrieves a specific user
      tags:
      - Users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: User retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  identification_number:
                    type: integer
                  balance:
                    type: number
                  pin:
                    type: integer
                  face_image_url:
                    type: string
        '404':
          description: User not found
    patch:
      summary: Updates a specific user
      tags:
      - Users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: User updated
        '404':
          description: User not found
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                identification_number:
                  type: integer
                balance:
                  type: number
                pin:
                  type: integer
                face_image_url:
                  type: string
    delete:
      summary: Deletes a specific user
      tags:
      - Users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: User deleted
        '404':
          description: User not found
servers:
- url: https://{defaultHost}
  variables:
    defaultHost:
      default: www.example.com